+++ package.json
@@ -44,8 +44,8 @@
     "typescript": "^3.3.4000"
   },
   "dependencies": {
-	"@kinecosystem/kin-sdk": "0.20.3",
-	"jest-each": "^24.7.1"
+    "@kinecosystem/kin-sdk": "0.20.3",
+    "jest-each": "^24.7.1"
   },
   "jest": {
     "moduleFileExtensions": [
